+++ client/react/src/App.js
@@ -13,12 +13,12 @@ import Plants from './pages/plant/Plants';
 import Home from './pages/Home';
 import Systems from './pages/system/Systems';
 import Alerts from './pages/alert/Alerts';
-import Todos from './pages/alert/todo/Todos';
+import Todos from './pages/todo/Todos';
 import Box from '@mui/material/Box';
 import AppNavigation from './AppNavigation';
 import NewPlantForm from './pages/plant/NewPlantForm';
 import NewSystemForm from './pages/system/NewSystemForm';
-import NewTodoForm from './pages/alert/todo/NewTodoForm';
+import NewTodoForm from './pages/todo/NewTodoForm';
 import NewTypeForm from './pages/plant/type/NewTypeForm';
 import NewLightForm from './pages/system/light/NewLightForm';
 import NewGenusForm from './pages/plant/genus/NewGenusForm';
